% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blip}
\alias{blip}
\title{Bayesian Learning Package - Main function.}
\usage{
blip(args)
}
\arguments{
\item{args}{Vector of arguments to be passed to the jar}
}
\description{
Used by most of the functions in the r.blip binding, provides access to the included jar file.
}
\details{
The arguments vector is formatted in a system call to the included jar file. 
Should not be called directly by the user, unless you know exactly what you are doing. 
In that case, call directly the blip jar.
}
