% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{blip.solver.tw}
\alias{blip.solver.tw}
\title{Structure Optimization - treewidth bound}
\usage{
blip.solver.tw(jkl, method = "kmax", treewidth = 4, time = 3600,
  cores = 1, verbose = 1)
}
\arguments{
\item{jkl}{cache of pre-computed parent sets.(required)}

\item{method}{Method to be used. Possible values: "kmax", "kg", "ka". (default: kmax)}

\item{treewidth}{Maximum treewidth (default: 4)}

\item{time}{Maximum Execution time (default: 3600)}

\item{cores}{Number of machine cores to use. If 0, all are used. (default: 1)}

\item{verbose}{Verbose level (default: 0)}
}
\value{
Structure
}
\description{
Find an optimal structure from the cache of parent sets
}
\details{
The input data is required to be complete and discrete. 
Accordingly missing values in the input data.frame will be ignored, and all numeric values will be converted to integers.
}
