\name{child}
\alias{child}
\docType{data}
\title{
  Child dataset
}
\description{
  Dataset generated from the famous "child" network.  Provided as an example of input data for learning a Bayesian network.
}
\usage{data("child")}
\format{
  The format is:
 chr "child"
}
\details{
 Space separeted, integer values for each variable.
}
\source{
 http://www.bnlearn.com/bnrepository/discrete-medium.html#child
}
\examples{
data(child)
}
\keyword{datasets}
