% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.jkl}
\alias{read.jkl}
\title{Jkl reader}
\usage{
read.jkl(path, names)
}
\arguments{
\item{path}{Path of the file to load}

\item{names}{List of variable names}
}
\value{
the cache of parent sets
}
\description{
Read a Jkl file (parent sets cache)
}
