% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.str}
\alias{read.str}
\title{Structure reader}
\usage{
read.str(path, names)
}
\arguments{
\item{path}{Path of the file to load}

\item{names}{List of variable names}
}
\value{
the BN structure
}
\description{
Reads a str file (BN structure)
}
