% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.jkl}
\alias{write.jkl}
\title{Jkl writer (with names)}
\usage{
write.jkl(path, jkl)
}
\arguments{
\item{path}{Path of the file to write}

\item{jkl}{parent sets cache to write}
}
\description{
Write a Jkl file (parent sets cache)
}
